+++ ubuntu_desktop_installer/lib/main.dart
@@ -4,11 +4,11 @@ import 'package:intl/intl.dart';
 import 'package:subiquity_client/subiquity_client.dart';
 import 'package:yaru/yaru.dart' as yaru;
 
-import 'keyboardlayoutpage.dart';
+import 'keyboard_layout_page.dart';
 import 'l10n/app_localizations.dart';
-import 'tryorinstallpage.dart';
-import 'turnoffrstpage.dart';
-import 'welcomepage.dart';
+import 'try_or_install_page.dart';
+import 'turn_off_rst_page.dart';
+import 'welcome_page.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
