+++ ubuntu_desktop_installer/lib/try_or_install_page.dart
